refactor(ai-tools): ai tools を packages に切り出した
