+++ apps/desktop/package.json
@@ -16,11 +16,10 @@
     "@faker-js/faker": "^10.3.0",
     "@repo/ai-chat": "workspace:*",
     "@repo/ai-chat-session": "workspace:*",
-    "@repo/async-queue": "workspace:*",
+    "@repo/ai-tools": "workspace:*",
     "@repo/auth": "workspace:*",
     "@repo/ipc": "workspace:*",
     "@repo/mcp-server-example": "workspace:*",
-    "@repo/rag": "workspace:*",
     "@repo/shadcn": "workspace:*",
     "@repo/sqlite": "workspace:*",
     "@repo/tanstack-ai-mcp": "workspace:*",
@@ -41,7 +40,6 @@
     "better-sqlite3": "^12.9.0",
     "express": "^5.2.1",
     "lucide-react": "^0.544.0",
-    "ollama": "^0.6.3",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-markdown": "^10.1.0",
