+++ packages/ai-tools/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@repo/ai-tools",
+  "version": "0.0.1",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./server/*": {
+      "types": "./src/server/*.ts",
+      "import": "./src/server/*.ts"
+    },
+    "./client/*": {
+      "types": "./src/client/*.ts",
+      "import": "./src/client/*.ts"
+    }
+  },
+  "scripts": {
+    "lint": "tsgo --noEmit && oxlint"
+  },
+  "dependencies": {
+    "@repo/rag": "workspace:*",
+    "@standard-schema/spec": "^1.1.0",
+    "@tanstack/ai": "^0.5.1",
+    "zod": "^4.3.6"
+  },
+  "devDependencies": {
+    "@types/node": "^24.12.2",
+    "@typescript/native-preview": "7.0.0-dev.20260418.1",
+    "oxlint": "^1.60.0",
+    "tsdown": "^0.21.10",
+    "typescript": "^6.0.3"
+  },
+  "peerDependencies": {
+    "electron": "^41.2.1"
+  },
+  "packageManager": "pnpm@10.33.0"
+}
