+++ packages/ai-tools/src/client/definitions.ts
