+++ packages/ai-tools/src/server/definitions.ts
